// PP-CoE-CodeApp/src/generated/models/PowerPlatformforAdminsModel.ts
export interface PolicyV2 {
  name: string;
  displayName: string;
  defaultConnectorsClassification: string;
  connectorGroups: {
   classification: string;
   connectors: {
    id: string;
    name: string;
    _type: string;
  }[];
 }[];
  environmentType: string;
  environments: {
   id: string;
   _type: string;
   name: string;
 }[];
  createdBy: {
    id?: string;
    displayName?: string;
  };
  createdTime: string;
  lastModifiedBy: {
    id?: string;
    displayName?: string;
  };
  lastModifiedTime: string;
  isLegacySchemaVersion: boolean;
}